Vimrc magic
I've added remappings to vim to automatically insert the matching reverse
bracket whenever a forward bracket or quote is typed.

I also added some conditional remapping to prevent inserting an extra reverse
bracket. If a reverse bracket is under the cursor, it will remap the bracket
to the right arrow so it appears to overtype the arrow instead of adding a
(probably redundant) reverse bracket.

Signed-off-by: Rick Henry <fredericmhenry@gmail.com>